[mem_cache][2/N] refactor: move SWATokenToKVPoolAllocator to allocator/swa.py, ancestry is recoverable via `git blame -C -C -C`.
